Change structure of the project. Main file divided into modules
